+++ activesupport/CHANGELOG.md
@@ -22,6 +22,10 @@
 
     *Jean Boussier*
 
+*   Add `ActiveSupport::Testing::InstrumentationAssertions` module to help with testing events.
+
+    *Nicholas La Roux*, *Yishu See*, *Sean Doyle*
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Include options when instrumenting `ActiveSupport::Cache::Store#delete` and `ActiveSupport::Cache::Store#delete_multi`.