+++ activesupport/CHANGELOG.md
@@ -10,6 +10,12 @@
 
     *Yasuo Honda*
 
+
+## Rails 6.1.7.9 (October 15, 2024) ##
+
+*   No changes.
+
+
 ## Rails 6.1.7.8 (June 04, 2024) ##
 
 *   No changes.